+++ src/App.css
@@ -68,7 +68,7 @@ body {
 }
 
 .heading {
-  font-size: 3rem;
+  font-size: 4rem;
   margin-bottom: 20px; 
   color: white; 
 }
